Update patch set 1
Patch Set 1: Code-Review+1

Hmm, that code was added specifically in a dedicated commit: 

That was 18 mnths ago so things might have changed (maybe it was needed despite not being "right"); have you checked with gryder@google.com, or could you? (if he's still at Google)

Other than that, I'm supportive of this change, as I too don't think aria-hidden is needed here; but this is just a gut feeling.

Patch-set: 1
Label: Code-Review=+1