Update patch set 1
Patch Set 1:

An interesting problem is whether to thread a CompilerContext and some of its components separately depending on the use.

IMHO either the CompilerContext is passed and all the needed components are extracted or only components are passed. And the  latter case should only happen at the bottom of the calls tree in functions that only need a few of those.

Patch-set: 1
Label: Code-Review=0