Update patch set 1
Patch Set 1:

The issue with the aria-hidden attribute is that a value of aria-hidden="false" is different than if the attribute applied at all.  This created issues where screen readers were reading child elements of a element with aria-hidden="false" even if they had had display:none explicitly set (due to how aria-hidden is designed).  If html5 hidden does not have those same issues, I would be fine with that.

Patch-set: 1
Label: Verified=0