Update patch set 1
Patch Set 1:

There may be those cases where you have a parent hidden (display:none and aria-hidden="true"), but the child element may need to be "visible" to screen readers (aria-hidden="false").  For example, text that needs to be read to the screen reader user to give descriptions/directions that aren't needed for visual users.

Patch-set: 1