Update patch set 1
Patch Set 1:

(8 comments)

Basically okay with it once the bug is fixed in CompilePerms.

Though it might take a while, at some point I'd like CompilerContext to go away in favor of dependency injection. A good step in that direction would be to pass the context to the constructor and store it as a field whenever possible, instead of passing it in each method call.

Patch-set: 1