Update patch set 1
Patch Set 1:

However, in that case shouldn't the aria-hidden="true" override the child?  According to w3.org, a value of true "Indicates that this section of the document and its children are hidden from the rendered view."  I honestly don't know which specificity is more important, the child or the parent.

Patch-set: 1