Update patch set 1
Patch Set 1:

It seems that we all agree that aria-hidden="true" should be set when setVisible(false) is called (I can make a change to this to update that).  However, due to the note on the W3 page advising not using aria-hidden="false", I recommend that we remove the attribute when setVisible(true) is called.  Would that be an acceptable solution?

Patch-set: 1