Update patch set 1
Patch Set 1:

Is it guaranteed that all screen readers will recognize that as a state transition?  Meaning, if you're going from aria-hidden="true" to the absence of that attribute (with the assumption it's visible) is it known to trigger the same state change as going from aria-hidden="true" to aria-hidden="false"?

Patch-set: 1