Update patch set 1
Patch Set 1:

I'm not an expert in this, so I don't know if it will be treated as a state transition.  But I do know that w3 indicates that you can either set it to false or remove it, so I would assume that they should treat both of those in the same way.

W3: "When the element is presented, authors MUST set the aria-hidden attribute to false or remove the attribute, indicating that the element is visible."

Patch-set: 1