Update patch set 1
Patch Set 1:

The idea would be to hide all the constructor arguments that can be supplied by the context without formally moving to dependency injection yet.

When using dependency injection, the framework calls the constructors and everything else calls through a Provider<T>.get() when there are no other arguments and factory interface if there are. So it would be a stepping stone. There are other intermediate points so that could be skipped if we decided on which Provider interface to use.

Patch-set: 1
Label: Code-Review=0