Update patch set 2
Patch Set 2:

i've applied roberto's "pass the CompilerContext or its components, but not both" rule throughout, except where the component being passed is a ResourceOracle (since resource oracles won't be populated into the CompilerContext till following patches)

i've also made CompilerContext completely immutable as brian suggested and i'm pretty happy with the change.

i have not yet worked on storing CompilerContext as an instance variable in preference over passing as a parameter, in preparation for switching to dependency injection. i'm going to do it but where it gets hairy because of static functions i'm going to leave the static functions alone.

in reference to "each class constructs other classes that need a CompilerContext by calling a factory method on the CompilerContext" if we're using a real dependency injection library (like Guice etc) then i don't think that is how we will implement it. but i could be wrong.

Patch-set: 2
Label: Verified=0