Update patch set 1
Patch Set 1:

The git rebase -i is OK yes; you could also use "git rebase -i master" (it could cause conflicts, but you'd have to rebase anyway later on, and in this case 'master' hasn't changed since your first commit). Then choose to "reword" the first commit and to "fixup" the second. And you should be ready to to push for review again, and finally "abandon" this review as it's no longer needed.

Patch-set: 1