Update patch set 3
Patch Set 3:

If it helps, I can explain why this fixes the issue -- since just the
one-character code change by itself isn't very descriptive I guess.

Say for example we have a south panel that's 200px tall with the splitter's
position at 500px, and the window is scrolled down by 100px.

When the mouse is pressed down, the class computes the "offset", which is the
mouse Y minus the splitter's Y. Since the mouse Y is in "screen" coordinates and
the splitter's Y is in "document" coordinates, this is nonzero when the window
is scrolled. In this case, the mouse would be at 400px on the screen, meaning
the offset is -100px.

Then let's say the mouse is dragged up by 10px. (So the new mouse Y is 390px.)
The current logic tries to set the new size:

new size = panel size + splitter Y - mouse Y - offset
         = 200px + 500px - 390px - (-100px)
         = 410px

This is wrong, since moving the mouse by 10px shouldn't change the panel size by
210px. By flipping the sign in the calculation from "- offset" to "+ offset", it
works out like this:

new size = panel size + splitter Y - mouse Y + offset
         = 200px + 500px - 390px + (-100px)
         = 210px

Which is the correct new size. Let me know if this makes sense.

Thanks

Patch-set: 3