Update patch set 3
Patch Set 3:

The code is shared with east splitter. Did you verify it was also broken and now fixed with the change as well?

Also it would be great if you can write test to re-produce the problem because it looks like that code path was never tested.

Patch-set: 3
Label: Verified=0