Update patch set 6
Patch Set 6:

I'm concerned that with this approach, essentially nothing is encapsulated, so we won't be able to change RequestData at all later without changing its public API. That seems counter-productive since the point is to merge the changes back later after experimenting.

What are the *minimal* changes that you need to create your subclass? If we can avoid adding a new public method then we should do so. Don't worry too much about reuse; for example, it should be okay to override tryInstall and onLoadError by copying them to the subclass and also copying all the private fields.

(Also, if we can get away with making something protected then we should do that instead of making it public.)

Patch-set: 6
Label: Verified=0