Fix for Type error nested try blocks followed by try block.
The cause of the error is a premature optimization in CatchBlockNormalizer.

CatchBlockNormalizer transforms the Java construct try-catch*-finally? into
the JavaScript try-catch-finally (only one catch block) introducing a variable
for each JS catch block. The optimization made sure that non nested try-catch
blocks would reuse the introduced variable.

However in GenerateJavaScriptAST each JsCatch block would get a new JsName
associated with its introduced variable, creating a situation where a
local has two corresponding JsName that could be obfuscated to different
names.

The last part of the puzzle is that the declaration of the introduced
variable in

  catch (var)

would use the JsName that was created in the catch scope wheres each
reference to the local variable would use the JsName associated
with the local variable resulting in incorrect code.

An example of the error is

  try { ...
  } catch (E e1) {
    try { ...
    } catch (E e2) {
    }
  }
  try { ...
  } catch ( E e3) {
  }

being translated to

  try { ...
  } catch (b) {
     a = ObfuscatedNameForUnWrap(a); // here it should have been b
     try { ...
     } catch (a) {
       a = ObfuscatedNameForUnWrap(a);
     }
  }
  try { ...
  } catch (a) {
     a = ObfuscatedNameForUnWrap(a);
  }

resulting in a type error when executing the first unwrap for a.

Fixes issue 7253.

Change-Id: I644e9e8dff569ec4b0431927b73b7fe0ffc83d40