Improves GWT event dispatch for extensibility.
New implementation introduces a dispatcher map for bitless events and
uses a 'default' dispatch function if there is no specific dispatcher
defined for the event.

Similarly, a dispatcher map for capture events introduced as well that
simplifies the boiler-plate around capture events.

With this change even the event is not known by GWT yet, it will be still
dispatched. As most events doesn't need custom dispatching, no further
changes will be required to DOMImpl (e.g. MGWT will no longer need to
rebind DOMImpl for transitionEnd and animationEnd as those will work out
of the box.

Change-Id: Ia00c177d0f007d544d4837e2291c991615795474