Update patch set 2
Patch Set 2:

I just rebased this on top of master. There are no semantic changes, I'm just fixing variable names that say "binary name" to be "internal name" in places where their values really are internal names.

This is part of my work from awhile to clean up source vs. internal vs. binary names in the compiler.

It is not super important, but would be nice to get some of the simpler parts of it in, I think.

Patch-set: 2
Label: Verified=0