Add a custom serializer for LinkedHashSet.
This allows LinkedHashSet to be serialized even if it's not in the serialization
policy file.

Unfortunately, several RPC tests used LinkedHashSet for testing boundary
conditions in generics handling, and these tests broke when the custom
field serializer was introduced. Which led to this commit getting reverted
previously.

Now we move the RPC tests to use a new "TestLinkedHashSet", which is a
copy of the existing LinkedHashSet, but still without a custom field
serializer, which allows the RPC tests to continue testing their boundary
conditions.

Change-Id: Ie230b4caf3de1b7a76a4c36324cba1f64ac8a1b3