Clarify JsPrettyNamer algorithm (take 2)
Introduced a changeNames() method to clarify the dataflow between
parent and child during recursion. Also, changed the return type of
JsScope.getAllNames() for convenience. Added tests for duplicate
detection in the same scope or a child scope.

(This is in preparation for adding a way to reduce the number of
globals.)

Change-Id: I5b60aff596f034df5fcc4294024b67713fc53a29