blacklist all globals in each browser
Before we weren't being comprehensive about this and just blacklisting
names we got from Closure and other places. Instead, let's use a script
to get all the globals in a browser and blacklist them whether we think
they're likely to be used or not.

This only includes Firefox and Chrome to start; please run globals.html
in whatever browsers you have and contribute more globals files.

(We probably need to do something else to make it less likely that there
will be a conflict with a browser or extension we don't know about at the
time of release, but that's out of scope for this patch.)

Change-Id: Ib526b006a30990a4f29c19b1011472278813ed18