Update patch set 1
Patch Set 1:

(3 comments)

This looks a lot like build dependencies where you redundantly declare what the code will do because it's too expensive to compute.

I think it's a fine idea but we need an automatic way to keep the declarations consistent with the code (sort of like strict deps), or they won't be kept to consistent and it will lead to hard-to-track-down compile errors.

An expensive way to do this would be to have a flag to turn on generator-checking mode, which runs the generator again (even though it shouldn't need to) and checks to see if the file contents changed. We would avoid setting this flag most of the time, but an app could set it in their continuous build to check for errors, or perhaps just when they suspect a problem.

Cheaper possibilities below.

Patch-set: 1