Update patch set 9
Patch Set 9:

Hey guys,

Not urgent, but I'm finally having time to think about this patch again. Things of note:

* The primary reason it's been a few months is that the IsXxx methods now all have javadocs that link to their concrete method. I thought these would be required for being accepted,  but they are also very boiler-platey, so I wanted to add them in an automated way. But I hadn't done stuff like that before, so pushed it to back of mind.

However, after some recent help from a friend, and a few fun nights playing with ANTLR, I was able auto-insert all of the comments. This was likely overkill, but it was fun, and means it would be easy to re-update them if we want them to be different (e.g. having a warning/experimental disclaimer, which I think would be a good idea.)

* Per Colin's ealier comments, this is a 90% effort; I would prefer to lazy-add interfaces for the rest of the widgets later, should the interfaces prove useful/in-demand to users.

* If these go in, I also have a whole suite of stub implementations that I would like to add next.

* Per Goktug's concerns, I know that a) we have next-gen widgets coming up, and b) MVP is not in vogue inside of Google anymore, but AFAIK the next-gen widgets will be separate (so that they can be completely modern), and there are a lot of people still doing MVP today that I think could incrementally start using these. And the stubs if/when those come.

Patch-set: 9
Label: Verified=0