Update patch set 12
Patch Set 12:

Patrick: no real reason for not using "@see". I did try it to verify that Eclipse hyperlinks it correctly, but for whatever reason I'm more used to @link.

Also, I did think that javadocs with no leading sentence, e.g. just an annotation (like "@see" or "@return") and that's it, don't seem to match the style of GWT's javadocs.

So then I thought of putting both a leading sentence + @see, but that seemed repetitive (but perhaps that's fine).

Perfectly happy to change it over/do both if you feel strongly about it.

Patch-set: 12