Makes LibraryBuilder path string encoding more forgiving.
When compiling actual applications Generator execution creates some
synthetic resources with an unusual path character (":"). Library
serialization was guarding against usage of this character, but it
has been changed to allow it. (This was originally correct but was
broken during out-of-context cleanup/review of the builder code).

Added a TODO to unify the Compiler on some fast and consistent
serialization library (probably protobuf) rather than the current
mixture of slow Java serialization, some protobuf, and custom
serialization.

Change-Id: Id4ef7477f3b81cf6a8b7c5eea6914a150d396fcf