Refactor JsoDevirtualizer in preparation for modular compilation.
JsoDevirtualizer is a pass that requires global knowledge.

For separate compilation we impose that the JSO must be declared in the
same module as the interface it implements. Also all JSOs are treated as
dualImpl JSOs in separate compilation.

This patch makes sure that all the required methods are defined even if not
called in the current module; the final pruner pass will remove static
methods that are not called

This patch should not have any effect for globally optimized applications.

Going forward devirtualizing JSO calls should be handled as all the other
normalization passes, and be run *before* optimizations; the optimizer
should be smart enough to produce code of similar quality.

Change-Id: Ic6457aab8a5889368a1ab538f7d26b3edc94d9c3