Update patch set 1
Patch Set 1:

(2 comments)

Roberto: "Is getIndent() guaranteed to be free of "-"? I am wondering if JsVerboseNamer would also need that fix."

i take it you mean that you noticed that the namers are updating name.setShortIdent() but never updating name.setIdent() and that maybe name.setIdent() should be updated as well? if that's what you meant, the answer is that the full ident can't be updated and only the short ident is used in javascript output after naming.

if that's not what you were asking, then i don't understand.

Patch-set: 1
Label: Verified=0