Adds a StatementRanges combining utility.
The Link system uses a StatementRanges object to know
how to and how much Javascript to write to disk. When writing
the combined Javascript from previously compiled libraries (as
part of separate compilation) it becomes necessary to combine
the StatementRanges for those libraries. Thus this change.

We want to enable separate compilation because it will be
faster than monolithic compilation.

Change-Id: I33879e213097c6335ea6c2369ad73dc274c6c8ea