Adds separate compilation flags and options.
Unlike the normal compiler entry point the separate compiler
entry point needs to be informed of precompiled input
libraries, the destination of its output library and whether
to perform linking (as linking is only performed in the last
separate compilation at the top of a compile tree).

So these options and args make it possible to drive the
separate compiler, which is deseriable since separate
compilation is faster.

Change-Id: I645f37d42b14cdecb3fd62e56cc395639a4e36bf
Review-Link: https://gwt-review.googlesource.com/#/c/5247/