Update patch set 2
Patch Set 2: Code-Review+1

Avoiding clever is good here. Not sure if this is the place, but JSOs and nulls are an interesting topic. Dev mode is of course sticky since any references to a null JSO end up hitting $wnd instead, so they work(ish) in dev mode, but can NPE in compiled JS (therefore SDM too). On the flip side, obj.cast() compiles out to obj, which is clearly null-safe, unless the compiler can prove that obj is null, in which case we get a null.nullMethod(). 

Again, not the place for it, but from watching this and https://gwt-review.googlesource.com/#/c/4360/, conservative changes in this dept are good.

Patch-set: 2
Reviewer: Gerrit User 5130 <5130@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=+1