Update patch set 4
Patch Set 4:

I don't think this is going to meet my midnight deadline for the 2.6 branch, but I think this is a relatively low risk change (only affects dev mode and testing, not production code).  If we can resolve the last few nits in the next day or two, I'll still approve cherry picking it for 2.6.  It's probably going to miss 2.6rc1 though. :(

Patch-set: 4
Label: Verified=0