Update patch set 2
Patch Set 2:

I decided to leave EmulationWithUserAgent.gwt.xml in there, even though it is currently empty. Some of the performance changes I am about to make require rebinding based on the user agent, so I am going to need the xml file.
Would it be okay to leave it in for now?

I am not very happy with the code duplication we got going on with StringBuilder and StringBuffer, but could not find a better way to handle this (since the type is embedded in the return type on every method). Introducing some kind of utility class might introduce problems with inlining.
Not sure about that.

Patch-set: 2