Update patch set 2
Patch Set 2:

Re: duplication - when this was first done, we tried three ways of sharing code.  This was the best assuming that only one would be used.  If you have a common implementation class and have the two implementations delegate to it, the compiler at the time would leave cruft and performance was worse as well.  It might be worth revisiting that approach and seeing if that is still true.

Patch-set: 2
Reviewer: Gerrit User 5100 <5100@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0