Update patch set 1
Patch Set 1: Code-Review+1

Hmm, guess I missed that it was merged over there. Ideally we wouldn't be *including* a built-in version of htmlunit that pretends to be FF, except when it doesn't (setAttributeNS, getElementsByTagNameNS).

For my own edification, what else gets tested in HtmlUnit, aside from the non-layout, non-htmlunit-is-buggy tests at Google? Should I just be giving up on htmlunit+devmode except for a simple set of smoke tests and moving any real testing to platform-specific builds of PhantomJS?

p.s. Tons of permgen space gets over that issue, and it does indeed appear that svg:image now has a setAttributeNS function, so images are working.

Patch-set: 1
Label: Code-Review=+1