GWT-RPC: don't inherit type variables for non-parameterized types
For example, suppose we're deserializing a field:

  Foo field

If Foo is a simple class or interface with no type parameters,
it shouldn't inherit the values of any type variables from its
environment. (The Foo class may have type variables of its own
but they should be independent.)

As a workaround I just made sure resolvedTypes is empty.
To do the job right probably requires a rewrite and public
API change for server-side custom serializers.

Includes a test.

Bug: issue 7628
Change-Id: Ib674812af6b8eabace42b4185ec5f07e967f1761