Update patch set 1
Patch Set 1:

Before trying to build, please note that it is necessary a new lib on $GWT_TOOLS! 
Check commit message for instruction.

This patch should change nothing to non-OSGi users, 
so the feedback expected is verification that no regression is introduced.

I'll provide some OSGi based GWT examples hopefully the this weekend, so that people interested into can test this patch on OSGi environment too.

In case of questions, write me here, on the comments to Issue 8424 or on the discussion on contributors mailing list.

Patch-set: 1
Label: Verified=0