Update patch set 2
Patch Set 2:

Regarding bnd - it can maintain package exports based on a whitelist as well, but does also some nice stuff like calculating imported packages and adding uses constraints, which are a pain to maintain manually. So I'd say that bnd still has its use.

Also, it can be used ( in the future ) to check that API changes are signalled properly in the exported package versions, which is also very valuable.

Patch-set: 2