Update patch set 2
Patch Set 2:

(I was posting this on the page of the issue but I see the discussion goes on here).

Usually it is very straightforward to manage bundles and it is recommended to follow as much as possible defaults of BND for import and export, but GWT-Servlet is a very difficult use case because: 
- it comes from java sources that does not get all compiled for the JVM 
- the Jar is also a merge and a filter of GWT-User and GWT-Dev
these two make it hard to configure properly BND, but it is feasible. We can use it to control the output and we don't need to write the header by hand.


Regarding the packages, I don't have the necessary knowledge to choose which package should be exported and which shouldn't. 
I want to avoid not-exporting a package required by someone else and that I'm not using.  

I see 2 approaches:
we either start from a wide set of exported packages and then reduce it in future releases of GWT,
or we start with a minimal set of exported packages and expand next time if required.

By myself alone I would leave the extra exports for "flexibility",
but if someone can help to enumerate the packages that need to be exported, we can adopt the conservative approach and I would prefer it too.  I will change the patch to reflect this list.

Anyhow I will also try to run significant GWT samples on Apache Karaf, this will allow me to refine and test the set of packages to be exported with more confidence.



Again, I prefer the conservative approach suggested by Robert, but I would prefer to include a first release in 2.6.0 eventually with more exports than what it is required if there is no time to follow the conservative approach for this release.

I think that a first release of OSGi-fied GWT-Servlet will allow people to start making tests and give their feedback and no one will 'tie' before next release (GWT 3.0 ?) which can be tuned with the right packages. 

In my job for example, we don't deploy (yet) GWT applications on OSGi, we use it only on the backend for implementing SOA's ESB, and we will not change this in GWT 2.6.0. But I would love to start a road that will make this feasible (and stable) in a near future, and I think it is needed a basic OSGi-fied GWT-Servlet.

In the end, also consider that if GWT 3.0 will follow a process of modularization, in next release it will probably be possible to fallback to using BND defaults (that could be also a proof of quality of modularization!).

Patch-set: 2