Update patch set 1
Patch Set 1:

Have you had a chance to make another pass at this (i.e. don't touch onunload, or wrap it, and change hosted.html as well)? I can put up my own proposal if you'd be able to test it out. It'd be very nice to announce at gwt.create that we've solved the memory leak once and for all ;).

Patch-set: 1