Update patch set 4
Patch Set 4:

I'm having problem with RPC serialization on OSGi.
The diagnosis is that something fails silently (with no exception) in OSGi when serializing a return value for a GWT RPC Service.

I have a GWT RPC service that returns a list of strings.

When I run it locally, this is the response (grabbed from the HTTP response):
//OK[6,2,5,2,4,2,3,2,4,1,["java.util.ArrayList/4159755760","java.lang.String/2004016611","string_1","string_2","string_3","string_4"],0,7]

When I run it in OSGi, the same example data is serialized to the following: 
//OK[4,1,["java.util.ArrayList/3047670833"],0,7]

the Server does not output any exception, but on the client side I get an error "The response could not be deserialized".

Anyone know any potential reason why GWT may fail to do serialization?

Patch-set: 4