In devmode, call plugin.disconnect() on window unload
This change hooks up the plugin.disconnect() call to the window
unload callbacks. (Notice that in both cases this actually hooks
them to an iframe window, not the parent window itself.)

If the plugin doesn't have a disconnect() method (true today) then it does nothing.

This is a prerequisite for fixing the Firefox memory leak.
For the next step, see:
https://gwt-review.googlesource.com/4680

Change-Id: Iff9f366bc7e34d5d2bb4098a49e0aaa3ca0aeb34