Update patch set 5
Patch Set 5:

Ignore last patch set, it is only after rebasing to the master.
I could have avoided to upload it but I had to know if it was working.

It (seem to me) has worked by using the command  'git rebase master', thank you Robert!

I try to clean and publish a first example. If I can't do it in time this evening,  I'll finish tomorrow.

Patch-set: 5
Label: Verified=0