Update patch set 5
Patch Set 5:

The second example I am trying on OSGi, the samples/validation, 
does not work due to a bug either on the "validation" example itself or on the gwt-servlet (the original one, not the one osgi-fied by me).

I'll send a mail to the contributors mailing list report it as here is off-topic.

But I want to ask you reviewer what do you expect me to achieve to try to accept this patch?


as far as now, I successfully make working a basic example created with webAppCreator (you can see it here: https://github.com/cristcost/gwt-karaf-examples )

and I'm trying to adapt to OSGi the following examples: validation, dynatable, dynatablerf, logexample, mobilewebapp (I have chosen only those with servlets in web.xml as it is only the server side code that make problems in OSGi)

I got stuck at example 1, validation, due to the intrinsic problem I will report into the mailing list. 
Would you please help me in addressing my next moves by discussing a set of condition to accept the patch?

I really thank you, as I care a lot to have this feature included in final 2.6.0, and I believe it could be of great benefit to all the GWT community.

Cristiano

Patch-set: 5