Update patch set 5
Patch Set 5:

for sure Robert,
but I think that as soon some other independent contributor like you, and Thomas will mark as +1 the patch, someone that can approve it may take care of this.
For now, this is only a patch that don't build due to the fact it does require a new library on GWT_TOOLS... I understand if other reviewer ignore it if they just see a red flag on gerrit!

...then yes I need your help to make it work then attract some attention :-)

anyway, thanks to the three of you a lot for the comments you have provided, the have really improved this contribute which suffers also from the fact it is the first time I use gerrit (and I'm almost a newbie on git...)

In the end, note that I also share your concerns and I believe this patch needs to be improved with more testing. I just want a chance from you to get your +1.

Cristiano

Patch-set: 5