Update patch set 4
Patch Set 4:

Oops, the issue is marked 2.6 but it looks like we abandoned it. Should we try to get it in? It seems relatively low-risk.

(The larger rationale: based on the issue tracker, some folks have an entirely custom but JDK-compatible logging framework with custom levels as well, so they need a constructor.)

Patch-set: 4