Update patch set 5
Patch Set 5:

Looks good but there is one issue though;

If the type corresponds to an event initialized twice (e.g. the type field is not marked as static in the event) then the behavior will be different and the handlers might be fired twice.

Not that I find a compelling reason to support this scenario, but I thought it might be a good idea to think&discuss the implications.

(I also don't like the idea of using a set instead of list here)

Patch-set: 5