Update patch set 4
Patch Set 4:

Not sure what you mean by "non op". I think the real issue is that in the JDK [1], Level doesn't have a no-argument constructor so there is no way to extend this class that compiles unless you use super-source. We could add a two-arg constructor matching the JDK that does nothing and then they could compile it without super-source, but it would be confusing that the arguments are ignored since that doesn't match the JDK's behavior, and we can't call it out in the javadoc in a way people are likely to see. (Also, perhaps their log framework is not GWT-specific code.)

Anyway, I would feel a little better dropping it since there's a workaround though it's kind of horrible.

[1] http://docs.oracle.com/javase/7/docs/api/java/util/logging/Level.html

Patch-set: 4
Reviewer: Gerrit User 5010 <5010@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0