Update patch set 5
Patch Set 5: Code-Review+1

(1 comment)

For the Set case: you can't even override hashcode because it is final. So you are right, there is no way that it can be useful.

It looks like event type is clearly intended to be used as singleton so I can't see anyway that would be useful with current implementation if it is not a singleton.

I'll wait one more day before +2 to see if anybody else has anything to add.

Thanks for the patch!

Patch-set: 5
Label: Code-Review=+1