Update patch set 1
Patch Set 1:

Yes, that's the idea. Basically we need a mechanism so that we can report exceptions to browser without changing the control flow.

I introduced maybeReportUncaughtException a few months back to guarantee the delivery of exceptions to testing infra but I wasn't fully happy with it because it was still not preventing some exceptions to get lost in production.

Patch-set: 1