Update patch set 2
Patch Set 2: -Code-Review

Hi,
all the applications I have tested worked, but I'have no idea of what can be using protobuf (for sure not my applications) and I've no idea if it is good to include it for the reason sated by Matthew or not...
I can just trust your experience regarding this.

My opinion is to go for the safe way, so with protobuf rebased included: its a 525K more on a 8M jar. Let's preserve the functionality.
Then in GWT 3.0 then a new organization of dependencies would be desirable.

I wait some word from you on which road to follow.

Patch-set: 2