Update patch set 1
Patch Set 1:

@Goktug

You can not choose Object.keys in JSNI directly since you need at least a polyfill (IE8 does not support it). 

I agree that we may not want to add this as a public API, since @JsInterface will be a better solution. How about I refactor this CL to go into core.client.impl and document this to be internal to the SDK. This will allow me to do the performance upgrades for emulation and other places, while not introducting a new public API. Once @JsInterface is in we can refactor to use it instead?

Patch-set: 1