Update patch set 1
Patch Set 1:

> You can not choose Object.keys in JSNI directly since 

If they care about IE8, it is trivial to check if Object.keys exists or not from JSNI - I'm not much concerned about external usage.

> How about I refactor this CL to go into core.client.impl and document this to be internal to the SDK.

That's reasonable but I think it would be better to start doing performance related changes first and then decide how to re-use the code for recurring patterns - we may not need a full polyfill just for our internal needs.

Patch-set: 1