Update patch set 1
Patch Set 1:

I think we will want to enable the new soyc reports using a flag at first and preserve the ability to do it the old way until everyone is confident in the new reports.

I haven't looked in detail, but one approach might be to put your changes into a SourceMapRecorder2 class and have the caller choose between them based on a flag. Is this feasible?

Patch-set: 1
Label: Verified=0