Update patch set 1
Patch Set 1:

The issue I see with this implementation is that it makes c.g.g.user depend on implementation details of c.g.g.resources, and more specifically, it makes it impossible or impractical for people that want to implement ImageResource themselves (like what people at Google are doing) to benefit from the "is standalone" behavior of the Image widget and ImageResourceRenderer.

I would care less if "Standalone" was an interface in the "public API" (outside the 'impl' subpackage), implemented by the ImageResourcePrototype.Standalone class (but that's an implementation detail).

On the other hand, this change is still an improvement compared to what's in GWT 2.5.1, so won't firmly oppose to it.

Patch-set: 1
Label: Code-Review=0