Update patch set 1
Patch Set 1:

All the internal users were implementing the interface for faking/mocking purposes. Some of them already using ImageResourcePrototype so they can mimic the standalone as well (for testing purposes).

Also we can go with ImageResourcePrototype.Bundle instead of ImageResourcePrototype.Standalone to make the default behavior closer to more common scenario.

I think the advantage of this implementation is, until someone asks for a way to benefit from "is standalone" behavior we can keep the api simple and compatible.

But if you think that would be certain then we can go with the API breaking change.

Patch-set: 1