Update patch set 2
Patch Set 2:

This addresses Brian's comments. When I try to respond to your comments through the UI, it fails.

For Firebug, I don't actually know the answer to that. When fully instrumented, Campaign Management fails to load in Firefox. I assume this is due to the size of the application.

Patch-set: 2
Label: Verified=0