Add runtime instrumentation support to GWT compiler.
1. New property "compiler.instrument" which toggles instrumentation
when value is "profile".
2. New configuration property "compiler.instrument.packagePrefix"
which optionally allows restriction of instrumentation to a package
sub-tree.
3. New Immortal Type, Instrumentation.java, which adds top level
utility functions instrument and profileFunction.
4. Altered GenerateJavaScriptAST. With instrumentation on, every
static method and virtual method is wrapped with a call to
profileFunction.
5. Instrumentation.instrument calls console.time at the start of the
method's body, and console.timeEnd at the end, providing timing
information to Chrome's chrome://tracing or Firebug.

Based on Ray Cromwell's original design and code, but adds
configuration property to allow limiting instrumentation to a
particular package subtree, and hard-codes the instrumentation to
call console.time.

Change-Id: Iec4b357a3632dcc9c25d5236703ec9b47197c6bb
Review-Link: https://gwt-review.googlesource.com/#/c/4800/