Update patch set 5
Patch Set 5:

Any thoughts on usability of this? For running animations, this is murder on chrome too, even with the inspector closed (apparently it records while not visible). Something like the $stats func to let you track without updating the UI, and a hint to not start recording until a certain time?

Along slightly different lines, it was discussed on a contrib call a month or so ago that this mechanism might be usable for coverage - is that still a reasonable proposal?

Patch-set: 5