Update patch set 5
Patch Set 5:

I think we can leave this one  

<T,U> T[] copyOfRange(U[] original, int from, int to, Class<? extends T[]> newType)

out as falling in the "GWT does not support reflection" umbrella.

(It would be possible to extend the emulated Class to support newInstance, but then optimization passes need to be aware of this to not prune the type, and it might end up blowing up the code).

Patch-set: 5
Reviewer: Gerrit User 5105 <5105@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0