Makes gwtOnLoad separate compilation compatible.
Existing linker and testing infrastructure expects GWT compiled
output to define exactly 1 gwtOnLoad bootstrap function. But
separate compilation needs to define a different gwtOnLoad for
every module.

This patch changes the construction of the gwtOnLoad bootstrap
function so that before it is defined a reference will be stored
to the previous gwtOnLoad and on execution it will first call
the previous one. As a result there is still only one top level
gwtOnLoad defined and available to linkers but its invocation
will chain invoke all of them that were defined. When compiling
monolithically there won't be a chain of gwtOnLoad calls but
when compiling separately there will be one per seperately
compiled module.

Change-Id: I6b5f0b70160fc22db0be95a5afa7c680b07c4e06
Review-Link: https://gwt-review.googlesource.com/#/c/5601/