Update patch set 1
Patch Set 1:

This patch has been standing for a while; I'm about to drop it but I would like to give it one more chance before that :)

As, I suggested earlier, I can add the "order" property so that the predicate execution order can be customized by third parties when a new permutation is added. e.g.

 <!-- Run IE11 predicate first (default order is '1' for all predicates) -->
 <set-configuration-property name="user.agent.ie11.order"
  value="0" />

@Colin: You are one of the primary customers for this; is this something you want to see in GWT?

Patch-set: 1