Fixes regression in GWTTestCase servlet handling
While we updated Jetty, instead of manually loading servlet with
WebAppContext.loadClass,  we stared using WebAppContext.addServlet
directly which loads the class internally.
Unfortunately, due to a bug in Jetty 8, calling addServlet with a
class that doesn't exist, leaves the Jetty in a broken state causing
any proceeding calls to addServlet fail.

This patch fixes that by rolling back the logic and let the class
load handled manually in JunitShell as before.

Change-Id: Ide48419f8c49c0c95f688d613041b96b8caf5f7a