Update patch set 1
Patch Set 1:

In GXT we find it necessary to add more user agents, but have been successful in adding our own property and using conditional properties to map back to the base 6 (now 5) that GWT itself has. From that perspective, we're pretty happy with the control we have. We can maintain our own set of user agents that are supported and ensure that our CssResources stay bound correctly to those property values - that was the main sticking point that led us to abandon making UserAgentPropertyGenerator easier to extend as you've done here. I originally threw together an approach in https://code.google.com/p/google-web-toolkit/issues/detail?id=6837, but even with that we found we were breaking things outside of our own sandbox, so it wasn't worth it.

So yes, I'd love to see this, and yes it would benefit us, but to be done right probably requires a bit more of a mindfulness toward this flexibility in adding/removing user agents. Between CssResource and the fallback spam, I'd be in favor of letting this go for now.

Patch-set: 1