Update patch set 2
Patch Set 2:

Okay, I think you have a good reason but ideally this would be a runtime setting. Maybe the "compile with -logLevel set to TRACE or DEBUG..." message in the log can be turned into a link that causes a recompile and redisplays the log?  That way people don't need to restart Super Dev Mode.

(There's no particular reason to hard-code. I just wanted to avoid overwhelming people with flags that aren't actually in use.)

Patch-set: 2