Added isDisplayed() convenience method
This is a method we came up with to deal with Issue 4892 in the tracker.  Just a convenience so end user does not need to crawl up the parent tree to see if their widget is actually visible in the browser.

Feel free to abandon if you don't think it fits with GWT or could cause confusion. Just wanted to try out the submission and review process. 

Change-Id: I9412f7a61d10b9dc12def18de747077f079fe21a